+++ github-updater.php
@@ -11,8 +11,8 @@
 /*
 Plugin Name:       GitHub Updater
 Plugin URI:        https://github.com/afragen/github-updater
-Description:       A plugin to automatically update GitHub or Bitbucket hosted plugins and themes into WordPress. Plugin class based upon <a href="https://github.com/codepress/github-plugin-updater">codepress/github-plugin-updater</a>. Theme class based upon <a href="https://github.com/WordPress-Phoenix/whitelabel-framework">Whitelabel Framework</a> modifications.
-Version:           4.2.0
+Description:       A plugin to automatically update GitHub or Bitbucket hosted plugins and themes. It also allows for remote installation of plugins or themes into WordPress. Plugin class based upon <a href="https://github.com/codepress/github-plugin-updater">codepress/github-plugin-updater</a>. Theme class based upon <a href="https://github.com/WordPress-Phoenix/whitelabel-framework">Whitelabel Framework</a> modifications.
+Version:           4.3.0
 Author:            Andy Fragen
 License:           GNU General Public License v2
 License URI:       http://www.gnu.org/licenses/gpl-2.0.html
@@ -24,31 +24,46 @@ Requires WP:       3.8
 Requires PHP:      5.3
 */
 
-// If this file is called directly, abort.
+/**
+ * Exit if called directly.
+ * PHP version check and exit.
+ */
 if ( ! defined( 'WPINC' ) ) {
 	die;
+}
+
+require_once ( plugin_dir_path( __FILE__ ) . '/vendor/WPUpdatePhp.php' );
+$updatePhp = new WPUpdatePhp( '5.3.0' );
+$updatePhp->set_plugin_name( 'GitHub Updater' );
+
+if ( ! $updatePhp->does_it_meet_required_php_version() ) {
+	return false;
 }
 
 // Load textdomain
 load_plugin_textdomain( 'github-updater', false, __DIR__ . '/languages' );
 
 // Plugin namespace root
-$root = array( 'Fragen\GitHub_Updater' => __DIR__ . '/src/GitHub_Updater' );
+$root = array( 'Fragen\\GitHub_Updater' => __DIR__ . '/src/GitHub_Updater' );
 
 // Add extra classes
-$extra_classes = array( 'Fragen\GitHub_Updater\Parsedown' => __DIR__ . '/vendor/Parsedown.php' );
+$extra_classes = array(
+	'Fragen\\GitHub_Updater\\Parsedown' => __DIR__ . '/vendor/Parsedown.php',
+	'WPUpdatePHP'                       => __DIR__ . '/vendor/WPUpdatePhp.php',
+	);
 
 // Load Autoloader
 require_once( __DIR__ . '/src/GitHub_Updater/Autoloader.php' );
-$class_loader = 'Fragen\GitHub_Updater\Autoloader';
-new $class_loader( $root, $extra_classes );
+$loader = 'Fragen\\GitHub_Updater\\Autoloader';
+new $loader( $root, $extra_classes );
 
 // Instantiate class GitHub_Updater
-new Fragen\GitHub_Updater\Base;
+$instantiate = 'Fragen\\GitHub_Updater\\Base';
+new $instantiate;
 
 /**
  * Calls Fragen\GitHub_Updater\Base::init() in init hook so other remote upgrader apps like
  * InfiniteWP, ManageWP, MainWP, and iThemes Sync will load and use all
  * of GitHub_Updater's methods, especially renaming.
  */
-add_action( 'init', array( 'Fragen\GitHub_Updater\Base', 'init' ) );
+add_action( 'init', array( 'Fragen\\GitHub_Updater\\Base', 'init' ) );